+++ moment/src/moment.ts
@@ -1,1 +1,2 @@
 export * from './modules/moment.module';
+export * from './misc/utils';